pyproject.toml
[tool.poetry]
name = "announcement-api"
version = "0.1.0"
description = ""
authors = ["Ivan Timofeev <timofeevis89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.4.0"
isort = "^5.13.2"
mypy = "^1.9.0"
flake8-pyproject = "^1.2.3"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migration
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 88
multi_line_output = 3
known_first_party = ["announcement_dj_pj"]

[tool.flake8]
max-line-length = 100
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    ".mypy_cache",
]
extend-ignore = "E203, E501"

[tool.mypy]
exclude = [".venv", "__pycache__", ".git"]
